+++ lib/puts_debuggerer/core_ext/kernel.rb
@@ -1,0 +1,6 @@
+module Kernel
+  def pd_inspect
+    pd self, printer: false, pd_inspect: true
+  end
+  alias pdi pd_inspect
+end