+++ spec/lib/puts_debuggerer_spec.rb
@@ -56,6 +56,13 @@ describe 'PutsDebuggerer' do
     expect(output).to eq("********************************************************************************\n[PD] #{puts_debuggerer_invoker_file}:56\n   > pd 'hello', 3, true, options\n  => [\"hello\", 3, true]\n********************************************************************************\n")
   end
 
+  xit 'prints using pd_inspect' do
+    result = PutsDebuggererInvoker.call_pd_inspect([1, [2, [3]]])
+    expect(result).to eq([1, [2, [3]]])
+    output = $stdout.string
+    expect(output).to eq("********************************************************************************\n[PD] #{puts_debuggerer_invoker_file}:56\n   > pd 'hello', 3, true, options\n  => [\"hello\", 3, true]\n********************************************************************************\n")
+  end
+
   context 'look into puts debuggerer blog post by tenderlove for other goodies to add'
   context 'deadlock detection support'
   context 'object allocation support' #might need to note having to load this lib first before others for this to work