+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Support `require 'pd`' as a shorter alternative to `require 'puts_debuggerer'`
 - Support `printer` as a Logger object or Logging::Logger (from "logging" gem). Basically any object that responds to :debug method.
 - Support `printer: false` option to return rendered String instead of printing and returning object
+- Set logger formatter to PutsDebuggerer::LOGGER_FORMATTER_DECORATOR when passing as printer (keeping format the same, but decorating msg with pd)
+- Add pd_inspect (and pdi alias) Kernel core extension methods
 
 ## 0.8.2